requirements.txt
absl-py==0.12.0
astunparse==1.6.3
cachetools==4.2.2
certifi==2021.5.30
chardet==4.0.0
cycler==0.10.0
decorator==4.4.2
easydict==1.9
gast==0.3.3
google-auth==1.31.0
google-auth-oauthlib==0.4.4
google-pasta==0.2.0
graphviz==0.8.4
grpcio==1.38.0
h5py==2.10.0
idna==2.10
imageio==2.9.0
joblib @ file:///tmp/build/80754af9/joblib_1613502643832/work
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
Markdown==3.3.4
matplotlib==3.4.2
mkl-fft==1.3.0
mkl-random @ file:///tmp/build/80754af9/mkl_random_1618853849286/work
mkl-service==2.3.0
mxnet-cu101==1.5.1
networkx==2.5.1
numpy==1.20.3
oauthlib==3.1.1
opencv-python==4.5.2.54
opt-einsum==3.3.0
Pillow==8.2.0
protobuf==3.17.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.7
python-dateutil==2.8.1
PyWavelets==1.1.1
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
scikit-image==0.18.1
scikit-learn @ file:///tmp/build/80754af9/scikit-learn_1621370412049/work
scipy @ file:///tmp/build/80754af9/scipy_1618855647378/work
six==1.16.0
tensorboard==2.2.2
tensorboard-plugin-wit==1.8.0
tensorflow-estimator==2.2.0
tensorflow-gpu==2.2.0
termcolor==1.1.0
threadpoolctl @ file:///tmp/tmp9twdgx9k/threadpoolctl-2.1.0-py3-none-any.whl
tifffile==2021.6.6
urllib3==1.26.5
Werkzeug==2.0.1
wrapt==1.12.1
